# Performance improvements for `shuffle` and `partial_shuffle`