--- a/benches/shuffle.rs
+++ b/benches/shuffle.rs
@@ -18,7 +18,6 @@ criterion_main!(benches);
 
 pub fn bench(c: &mut Criterion) {
     bench_rng::<rand_chacha::ChaCha12Rng>(c, "ChaCha12");
-    bench_rng::<rand_chacha::ChaCha20Rng>(c, "ChaCha20");
     bench_rng::<rand_pcg::Pcg32>(c, "Pcg32");
     bench_rng::<rand_pcg::Pcg64>(c, "Pcg64");
 }