--- a/src/seq/increasing_uniform.rs
+++ b/src/seq/increasing_uniform.rs
@@ -79,9 +79,9 @@ impl<R: RngCore> IncreasingUniform<R> {
 #[inline]
 /// Calculates `bound`, `count` such that bound (m)*(m+1)*..*(m + remaining - 1)
 fn calculate_bound_u32(m: u32) -> (u32, u8) {
+    debug_assert!(m > 0);
     #[inline]
     const fn inner(m: u32) -> (u32, u8) {
-        debug_assert!(m > 0);
         let mut product = m;
         let mut current = m + 1;