This is Related to https://github.com/rust-random/rand/issues/1266 but completely orthogonal to https://github.com/rust-random/rand/pull/1268 which improves the performance of a different set of methods in a different way.

This improves the performance of `SliceRandom::shuffle()` and `SliceRandom::partial_shuffle()` by essentially batching the random number generation.

It seems to be about 50-100% faster for most slice lengths, with less performance improvement for longer slices. It will use the old method for slices with length longer than 2^32.

This is a value breaking change.

Benchmark results

## Partial Shuffle
Partial shuffle half of the slice
| Number of Elements | Rng       | Old ns/iter | New ns/iter | Ratio (new:old) |
| ------------------ | --------- | ----------- | ----------- | --------------- |
| 10                 | CryptoRng | 44          | 27          | 1.62962963      |
| 10                 | SmallRng  | 30          | 21          | 1.428571429     |
| 100                | CryptoRng | 409         | 164         | 2.493902439     |
| 100                | SmallRng  | 270         | 130         | 2.076923077     |
| 1000               | CryptoRng | 3,616       | 1,993       | 1.814350226     |
| 1000               | SmallRng  | 2,474       | 1,364       | 1.813782991     |
| 10000              | CryptoRng | 38,607      | 26,286      | 1.468728601     |
| 10000              | SmallRng  | 27,824      | 16,927      | 1.6437644       |

## Shuffle
| Number of Elements | Rng       | Old ns/iter | New ns/iter | Ratio (new:old) |
| ------------------ | --------- | ----------- | ----------- | --------------- |
| 1                  | CryptoRng | 0           | 0           | N/A             |
| 1                  | SmallRng  | 0           | 0           | N/A             |
| 2                  | CryptoRng | 11          | 8           | 1.375           |
| 2                  | SmallRng  | 8           | 5           | 1.6             |
| 3                  | CryptoRng | 18          | 9           | 2               |
| 3                  | SmallRng  | 13          | 6           | 2.166666667     |
| 10                 | CryptoRng | 88          | 25          | 3.52            |
| 10                 | SmallRng  | 61          | 16          | 3.8125          |
| 100                | CryptoRng | 872         | 325         | 2.683076923     |
| 100                | SmallRng  | 552         | 245         | 2.253061224     |
| 1000               | CryptoRng | 7,219       | 3,910       | 1.84629156      |
| 1000               | SmallRng  | 5,057       | 2,650       | 1.908301887     |
| 10000              | CryptoRng | 76,061      | 50,041      | 1.519973622     |
| 10000              | SmallRng  | 55,682      | 32,715      | 1.702032707     |