@TheIronBorn 2022-12-06
Huh, I've always thought shuffling was memory bound. Impressive work