@wainwrightmark 2023-01-04
If we did have a way of determining native output size (https://github.com/rust-random/rand/issues/1261) using 64 bit chunks would give significant performance improvements when shuffling longer sequences.  Unfortunately this would lead to different values on 32 and 64 bit machines.