@dhardy 2023-01-04
>  Unfortunately this would lead to different values on 32 and 64 bit machines.

Only where a different RNG is used on these machines (e.g. `SmallRng` which is a platform-dependent type-def), in which case results would already differ.

There is a question of which chunk size we should use by default given that 64-bit CPUs are now the norm, though it would penalise results for short lists with e.g. ChaCha.

Actually, we should run benchmarks with both chunk sizes with several RNGs (e.g. Pcg32, Pcg64 and ChaCha12; don't need two variants of ChaCha) — as a hack that doesn't need to be committed. I'll do this for your other PR.