@wainwrightmark 2023-01-04
I ran the benchmarks comparing u32 and u64 versions.  
Unsurprisingly, the main factor seems to be the number of elements. Regardless of the RNG, with 10 or fewer elements using u64 is a lot slower, with 100 it's about the same, and with 1000 or 10000 you start to see the benefit.   
Of course this is me running on a 64 bit machine - on a 32 bit machine all those 64 bit div operations would probably be a lot slower.


|Method         |Elements|Rng     |units|u32   |u64   |Ratio     |
|---------------|--------|--------|-----|------|------|------------|
|shuffle        |1       |ChaCha12|ps   |147.86|161   |1.088867848 |
|shuffle        |2       |ChaCha12|ns   |7.6755|17.923|2.335092176 |
|shuffle        |3       |ChaCha12|ns   |9.0967|19.445|2.137588356 |
|shuffle        |10      |ChaCha12|ns   |19.83 |43.989|2.218305598 |
|partial_shuffle|10      |ChaCha12|ns   |22.278|37.443|1.680716402 |
|shuffle        |100     |ChaCha12|ns   |296   |310.14|1.04777027  |
|partial_shuffle|100     |ChaCha12|ns   |153.41|164.68|1.073463268 |
|shuffle        |1000    |ChaCha12|µs   |3.3972|3.239 |0.9534322383|
|partial_shuffle|1000    |ChaCha12|µs   |1.7698|1.5642|0.8838286812|
|shuffle        |10000   |ChaCha12|µs   |42.689|35.48 |0.8311274567|
|partial_shuffle|10000   |ChaCha12|µs   |22.199|17.56 |0.7910266228|
|shuffle        |1       |Pcg32   |ps   |153.99|158.82|1.031365673 |
|shuffle        |2       |Pcg32   |ns   |6.2062|13.458|2.168476685 |
|shuffle        |3       |Pcg32   |ns   |7.2483|15.499|2.138294497 |
|shuffle        |10      |Pcg32   |ns   |16.559|38.7  |2.337097651 |
|partial_shuffle|10      |Pcg32   |ns   |29.098|33.407|1.148085779 |
|shuffle        |100     |Pcg32   |ns   |238.94|281.99|1.180170754 |
|partial_shuffle|100     |Pcg32   |ns   |128.43|147.25|1.146538971 |
|shuffle        |1000    |Pcg32   |µs   |2.667 |2.6314|0.9866516685|
|partial_shuffle|1000    |Pcg32   |µs   |1.3933|1.2401|0.8900452164|
|shuffle        |10000   |Pcg32   |µs   |32.283|28.93 |0.8961372859|
|partial_shuffle|10000   |Pcg32   |µs   |16.62 |14.075|0.8468712395|
|shuffle        |1       |Pcg64   |ps   |176.42|149.3 |0.8462759324|
|shuffle        |2       |Pcg64   |ns   |7.4931|13.865|1.850369006 |
|shuffle        |3       |Pcg64   |ns   |8.5974|15.291|1.77856096  |
|shuffle        |10      |Pcg64   |ns   |19.851|37.843|1.906352325 |
|partial_shuffle|10      |Pcg64   |ns   |21.434|30.914|1.442287954 |
|shuffle        |100     |Pcg64   |ns   |265.08|291.27|1.098800362 |
|partial_shuffle|100     |Pcg64   |ns   |141.66|151.39|1.068685585 |
|shuffle        |1000    |Pcg64   |µs   |3.0694|2.7147|0.8844399557|
|partial_shuffle|1000    |Pcg64   |µs   |1.6057|1.3093|0.8154076104|
|shuffle        |10000   |Pcg64   |µs   |39.199|29.133|0.7432077349|
|partial_shuffle|10000   |Pcg64   |µs   |20.264|14.098|0.6957165417|