@wainwrightmark 2023-01-05
> `Pcg64` being faster with 64-bit chunks for large sizes is not surprising since the 32-bit version is discarding random bits, but the significant losses for ≤ 10 elements and only moderate wins at 10'000 elements means it's still questionable whether 64-bit chunks is an improvement.
> 
> Meanwhile ChaCha and Pcg32 also see gains despite not discarding random bits in the same way.
> 
> We _could_ use this to select a different shuffling implementation based on the slice length, regardless of RNG algorithm. But is there enough interest in shuffling large slices to justify the extra complexity? Probably better to stick with 32-bit only.
> 
> Thanks for the extra benchmarks @wainwrightmark.

I agree.